Minor: move code around a bit, drop some unused code
