+++ src/main/java/org/alessio29/savagebot/bennies/Benny.java
@@ -1,4 +1,4 @@
-package org.alessio29.savagebot.savagebot.bennies;
+package org.alessio29.savagebot.bennies;
 
 public class Benny {
