+++ src/main/java/org/alessio29/savagebot/commands/bennies/UseBennyCommand.java
@@ -8,9 +8,9 @@ import org.alessio29.savagebot.commands.Category;
 import org.alessio29.savagebot.commands.ICommand;
 import org.alessio29.savagebot.internal.CommandExecutionResult;
 import org.alessio29.savagebot.internal.Messages;
-import org.alessio29.savagebot.savagebot.bennies.BennyColor;
-import org.alessio29.savagebot.savagebot.bennies.Pocket;
-import org.alessio29.savagebot.savagebot.bennies.Pockets;
+import org.alessio29.savagebot.bennies.BennyColor;
+import org.alessio29.savagebot.bennies.Pocket;
+import org.alessio29.savagebot.bennies.Pockets;
 
 
 public class UseBennyCommand implements ICommand {
