--- src/main/java/org/alessio29/savagebot/r2/EvaluationResultsListener.java
@@ -1,7 +1,0 @@
-package org.alessio29.savagebot.r2;
-
-import org.alessio29.savagebot.r2.tree.Expression;
-
-interface EvaluationResultsListener {
-    void putExplanation(Expression expression, String explanation);
-}
