+++ src/main/java/org/alessio29/savagebot/r2/eval/EvaluationErrorException.java
@@ -1,4 +1,4 @@
-package org.alessio29.savagebot.r2;
+package org.alessio29.savagebot.r2.eval;
 
 class EvaluationErrorException extends RuntimeException {
     EvaluationErrorException(String s) {
