+++ src/main/java/org/alessio29/savagebot/r2/eval/ExpressionContext.java
@@ -1,9 +1,8 @@
-package org.alessio29.savagebot.r2;
+package org.alessio29.savagebot.r2.eval;
 
 import org.alessio29.savagebot.r2.tree.Expression;
 
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 
 class ExpressionContext {
