+++ src/main/java/org/alessio29/savagebot/r2/eval/IntResult.java
@@ -1,4 +1,4 @@
-package org.alessio29.savagebot.r2;
+package org.alessio29.savagebot.r2.eval;
 
 import java.util.Comparator;
