+++ src/main/java/org/alessio29/savagebot/r2/eval/Interpreter.java
@@ -1,6 +1,7 @@
-package org.alessio29.savagebot.r2;
+package org.alessio29.savagebot.r2.eval;
 
 import org.alessio29.savagebot.internal.Messages;
+import org.alessio29.savagebot.r2.Dumper;
 import org.alessio29.savagebot.r2.tree.NonParsedStringStatement;
 import org.alessio29.savagebot.r2.tree.Statement;
