+++ src/test/java/tests/TestR2Parser.java
@@ -1,7 +1,7 @@
 package tests;
 
 import org.alessio29.savagebot.r2.Dumper;
-import org.alessio29.savagebot.r2.Parser;
+import org.alessio29.savagebot.r2.parse.Parser;
 import org.alessio29.savagebot.r2.tree.Statement;
 import org.junit.Assert;
 import org.junit.Test;
